refactor : 보드 패키지 이동으로 인한 import문 변경
